+++ project/test/isolated/test_limit_repeating_error_window.gd
@@ -26,10 +26,15 @@ func _before_send(_ev: SentryEvent) -> SentryEvent:
 
 ## Only one error should be logged within 1 second time window, and another one after 1 second passes.
 func test_repeating_error_window_limit() -> void:
-	var monitor := monitor_signals(self, false)
+	# Wait for special startup limits to expire.
+	while Engine.get_process_frames() < 10:
+		await get_tree().process_frame
+
+	monitor_signals(self, false)
+
 	push_error("dummy-error")
 	push_error("dummy-error")
-	await assert_signal(monitor).is_emitted("callback_processed")
+	await assert_signal(self).is_emitted("callback_processed")
 	assert_int(_num_events).is_equal(1)
 
 	# Wait for window to expire.
@@ -37,5 +42,5 @@ func test_repeating_error_window_limit() -> void:
 
 	push_error("dummy-error")
 	push_error("dummy-error")
-	await assert_signal(monitor).is_emitted("callback_processed")
+	await assert_signal(self).is_emitted("callback_processed")
 	assert_int(_num_events).is_equal(2)