+++ src/sentry/android/android_string_names.cpp
@@ -31,6 +31,7 @@ AndroidStringNames::AndroidStringNames() {
 	createEvent = StringName("createEvent");
 	releaseEvent = StringName("releaseEvent");
 	captureEvent = StringName("captureEvent");
+	captureFeedback = StringName("captureFeedback");
 	addFileAttachment = StringName("addFileAttachment");
 	addBytesAttachment = StringName("addBytesAttachment");
 
@@ -63,6 +64,9 @@ AndroidStringNames::AndroidStringNames() {
 	releaseException = StringName("releaseException");
 	exceptionAppendStackFrame = StringName("exceptionAppendStackFrame");
 	eventAddException = StringName("eventAddException");
+	eventGetExceptionCount = StringName("eventGetExceptionCount");
+	eventSetExceptionValue = StringName("eventSetExceptionValue");
+	eventGetExceptionValue = StringName("eventGetExceptionValue");
 
 	// Breadcrumbs.
 	createBreadcrumb = StringName("createBreadcrumb");