+++ src/sentry/android/android_string_names.h
@@ -43,6 +43,7 @@ public:
 	StringName createEvent;
 	StringName releaseEvent;
 	StringName captureEvent;
+	StringName captureFeedback;
 	StringName addFileAttachment;
 	StringName addBytesAttachment;
 
@@ -75,6 +76,9 @@ public:
 	StringName releaseException;
 	StringName exceptionAppendStackFrame;
 	StringName eventAddException;
+	StringName eventGetExceptionCount;
+	StringName eventSetExceptionValue;
+	StringName eventGetExceptionValue;
 
 	// Breadcrumbs.
 	StringName createBreadcrumb;